Initial version of Levo.ai's Burp extension
The extension allows users to build OpenApi specifications for
their APIs based on the traffic from Burp's browsing.
The traffic is sent from the extension to Levo's Satellite service
that's expected to run locally at 9999 port.

In future versions, we plan to eliminate the dependency on
running Levo's services locally